Combined ILM changes and Last Tweak Changes
ILM changes (from Nick): more build and warning fixes, for separate build
roots and some other minor cleanups (like successfully running make
distcheck).

random access reading support for properties and components.

fixed a bug in the ostream based writer where it was attempting to
free the stream that was passed in.

Added ordering support to the string table to prevent string attribute
indices from having to be rewritten when written to the file.

Tweak changes include GTO_ prefix on GTO Parser.y tokens

+++ lib/Gto/RawData.cpp
@@ -347,7 +347,7 @@ RawDataBaseWriter::write(const char *filename,
 
             const Components &components = o.components;
 
-            for (int q=0; q < components.size(); q++)
+            for (size_t q=0; q < components.size(); q++)
             {
                 writeComponent(true, components[q]);
             }
